requirements.txt
flask
Werkzeug
kiteconnect
python-dotenv
psutil
psycopg2-binary
pytz
pyopenssl
requests

pandas
numpy
APScheduler
croniter
smartapi-python
pyotp
logzero
websocket-client
